# metal : refactor kernel args into structs