--- a/ggml/src/ggml-common.h
+++ b/ggml/src/ggml-common.h
@@ -514,6 +514,24 @@ typedef struct {
     int32_t  nei1;
     uint64_t nbi1;
     int32_t  ne00;
+    int32_t  ne02;
+    uint64_t nb01;
+    uint64_t nb02;
+    int32_t  ne11;
+    int32_t  ne12;
+    int32_t  ne13;
+    uint64_t nb10;
+    uint64_t nb11;
+    uint64_t nb12;
+    int32_t  ne0;
+    int32_t  ne1;
+} ggml_metal_kargs_mul_mm_id;
+
+typedef struct {
+    int32_t  nei0;
+    int32_t  nei1;
+    uint64_t nbi1;
+    int32_t  ne00;
     int32_t  ne01;
     int32_t  ne02;
     uint64_t nb00;