--- a/ggml/src/ggml-common.h
+++ b/ggml/src/ggml-common.h
@@ -495,6 +495,25 @@ typedef struct {
 } ggml_metal_kargs_repeat;
 
 typedef struct {
+    int64_t  ne00;
+    int64_t  ne01;
+    int64_t  ne02;
+    int64_t  ne03;
+    uint64_t nb00;
+    uint64_t nb01;
+    uint64_t nb02;
+    uint64_t nb03;
+    int64_t  ne0;
+    int64_t  ne1;
+    int64_t  ne2;
+    int64_t  ne3;
+    uint64_t nb0;
+    uint64_t nb1;
+    uint64_t nb2;
+    uint64_t nb3;
+} ggml_metal_kargs_cpy;
+
+typedef struct {
     int32_t  ne00;
     int32_t  ne01;
     int32_t  ne02;