… ggml/src/ggml-metal/CMakeLists.txt
… ggml/src/ggml-metal/ggml-metal.metal